Remove foreign key constraint on file_status table
There is a race condition between the add file endpoint and the file status endpoint. If the file is not in the DB when the first status message comes back we get a foreign key constraint error. We could fix this, but it would be expensive and not really worth the effort